+++ quimb/tensor/tensor_core.py
@@ -4337,9 +4337,16 @@ class TensorNetwork(object):
         """The total number of indices in the tensor network."""
         return len(self.ind_map)
 
-    def pop_tensor(self, tid):
+    def pop_tensor(self, tid_or_tags, which="all") -> Tensor:
         """Remove tensor with ``tid`` from this network, and return it."""
         # pop the tensor itself
+
+        if isinstance(tid_or_tags, int):
+            tid = tid_or_tags
+        else:
+            # get the tensor ids from the tags
+            tid, = self._get_tids_from_tags(tid_or_tags, which=which)
+
         t = self.tensor_map.pop(tid)
 
         # remove the tid from the tag and ind maps