+++ docs/changelog.md
@@ -27,6 +27,7 @@ Release notes for `quimb`.
 - [`tensor_split`](quimb.tensor.tensor_core.tensor_split): add `matrix_svals` option, if `True` any returned singular values are put into the diagonal of a matrix (by default, `False`, they are returned as a vector).
 - add [`Tensor.new_ind_pair_diag`](quimb.tensor.tensor_core.Tensor.new_ind_pair_diag) for expanding an existing index into a pair of new indices, such that the diagonal of the new tensor on those indices is the old tensor.
 - [`TNOptimizer`](quimb.tensor.optimize.TNOptimizer): add 'cautious' ADAM
+- [`TensorNetwork.pop_tensor`](quimb.tensor.tensor_core.TensorNetwork.pop_tensor): allow `tid` or tags to be specified.
 
 **Bug fixes:**